refactor: split store.rs tests into concern-based modules (issue #358)
Move the ~2,600 lines of #[cfg(test)] content out of src/outbound/sql/store.rs
into store/tests/{fixtures,credentials,status_list,history,contention}.rs,
consolidating duplicate sqlite/mysql/postgres tests into shared assert_*
helpers. No production code changed; the implementation stays in store.rs
with a thin #[cfg(test)] mod tests; re-export.
